requirements.txt
astroid==2.12.13
async-generator==1.10
attrs==22.1.0
autopep8==2.0.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
dill==0.3.6
exceptiongroup==1.0.4
google-api-core==2.8.2
google-api-python-client==2.51.0
google-auth==2.8.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
gspread==5.7.1
h11==0.14.0
httplib2==0.21.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.2.0
packaging==21.3
platformdirs==2.5.4
pluggy==1.0.0
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.2
pylint==2.15.6
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
webdriver-manager==3.8.5
wrapt==1.14.1
wsproto==1.2.0
